+++ Pug-Frontend/src/app/services/auth.service.ts
@@ -3,35 +3,28 @@ import { CookieService } from 'ngx-cookie-service';
 import { HttpClient } from '@angular/common/http';
 import { environment } from 'environments/environment';
 import { NotificationsService } from './notifications.service';
+import { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { Observable } from 'rxjs';
 
 
 @Injectable({
   providedIn: 'root'
 })
-export class AuthService {
+export class AuthService implements CanActivate{
 
   constructor(
-    private cookies: CookieService, 
+    private cookies: CookieService,
     private http: HttpClient,
     private notificationsService: NotificationsService
     ) { }
 
-  public async login(username:string, password:string) {
-    this.http.post(environment.url + 'auth/login/', { username, password }).subscribe(
-          async (token) => {
-              console.log(token);
-              this.setToken(token);
-          },
-          async (error) => {
-              this.notificationsService.showNotification('Las credenciales son incorrectas', true);
-          }
-      );
+  // tslint:disable-next-line: max-line-length
+  canActivate(
+    _route: ActivatedRouteSnapshot,
+    _state: RouterStateSnapshot
+    ): boolean | UrlTree | Observable<boolean | UrlTree> | Promise<boolean | UrlTree> {
+      
+    throw new Error('Method not implemented.');
   }
 
-  private setToken(token: any) {
-    this.cookies.set("token", token);
-  }
-  getToken() {
-    return this.cookies.get("token");
-  }
 }
